refactor: #265: split route book into multiple files and improve styles
