+++ src/pages/trade/ui/route-book/index.ts
@@ -1,0 +1,1 @@
+export { RouteBook } from './book';
